Create modelstest package
I want to be able to use an in memory database in my testing for the main CLI, so I made that function part of its own package. I had to swap over the internal testing to its own package to avoid an import cycle. I like this anyways because it shows how my code is consumed.