Big refactor!
So I wanted to give a more descriptive package name than just "models". I chose "transaction" because it reduced a lot of the constants and I found that it very logically grouped the code in the package. I also realized I could reduce my API surface, so I did that too. I still have to go back and readd the tests though...