Add ability to see the current month's spending in the recent command.
This entailed adding the Range method to the Table interface. I am considering adding a RangeTotal, because that would match more closely what I'm doing. I figured the cost for getting the range is low for going a month back at most. We'll see.

I also removed the limit from ScanSet. If somebody uses it in the wrong way, that's on them.